+++ packages/transaction-pay-controller/src/strategy/relay/polymarket/calldata.ts
@@ -1,9 +1,15 @@
+import { Interface } from '@ethersproject/abi';
 import type { Hex } from '@metamask/utils';
 
-const ERC20_TRANSFER_SELECTOR = '0xa9059cbb';
+const iface = new Interface([
+  'function approve(address spender, uint256 amount)',
+  'function unwrap(address asset, address recipient, uint256 amount)',
+  'function wrap(address asset, address recipient, uint256 amount)',
+  'function transfer(address recipient, uint256 amount)',
+]);
 
 export function encodeApprove(spender: Hex, amount: bigint): Hex {
-  return `0x095ea7b3${padAddress(spender)}${padUint256(amount)}` as Hex;
+  return iface.encodeFunctionData('approve', [spender, amount]) as Hex;
 }
 
 export function encodeUnwrap({
@@ -15,7 +21,7 @@ export function encodeUnwrap({
   recipient: Hex;
   amount: bigint;
 }): Hex {
-  return `0x8cc7104f${padAddress(asset)}${padAddress(recipient)}${padUint256(amount)}` as Hex;
+  return iface.encodeFunctionData('unwrap', [asset, recipient, amount]) as Hex;
 }
 
 export function encodeWrap({
@@ -27,22 +33,10 @@ export function encodeWrap({
   recipient: Hex;
   amount: bigint;
 }): Hex {
-  return `0x62355638${padAddress(asset)}${padAddress(recipient)}${padUint256(amount)}` as Hex;
+  return iface.encodeFunctionData('wrap', [asset, recipient, amount]) as Hex;
 }
 
 export function extractErc20TransferRecipient(data: Hex): Hex {
-  if (!data.startsWith(ERC20_TRANSFER_SELECTOR)) {
-    throw new Error(
-      `Expected ERC-20 transfer calldata, got selector ${data.slice(0, 10)}`,
-    );
-  }
-  return `0x${data.slice(34, 74)}` as Hex;
-}
-
-function padAddress(address: Hex): string {
-  return address.slice(2).toLowerCase().padStart(64, '0');
-}
-
-function padUint256(value: bigint): string {
-  return value.toString(16).padStart(64, '0');
+  const [recipient] = iface.decodeFunctionData('transfer', data);
+  return recipient as Hex;
 }
